refactor(core): extract FFI input decoding
